Move TestRPMDistGit tests to dedicated file
